Extract the run() method into an AbstractTokenizer
The method is the same in every implementation so we should move it up
into a base class to inherit from.
