+++ src/Tokenizer/PunctuationTokenizer.php
@@ -34,7 +34,7 @@
 namespace Org\Heigl\Hyphenator\Tokenizer;
 
 /**
- * Use Punktuation to split any input into tokens
+ * Use Punctuation to split any input into tokens
  *
  * @category   Hyphenation
  * @package    Org_Heigl_Hyphenator
@@ -46,9 +46,8 @@ namespace Org\Heigl\Hyphenator\Tokenizer;
  * @link       http://github.com/heiglandreas/Hyphenator
  * @since      04.11.2011
  */
-class PunctuationTokenizer implements Tokenizer
+class PunctuationTokenizer extends AbstractTokenizer
 {
-
     /**
      * The tokens to be handled by this tokenizer as an array.
      *
@@ -96,56 +95,16 @@ class PunctuationTokenizer implements Tokenizer
     );
 
     /**
-     * Split the given input into tokens using punktuation marks as splitter
+     * Split the given input into tokens using punctuation marks as splitter.
      *
-     * The input can be a string or a tokenRegistry. If the input is a
-     * TokenRegistry, each item will be tokenized.
-     *
-     * @param string|\Org\Heigl\Hyphenator\Tokenizer\TokenRegistry $input The
-     * input to be tokenized
-     *
-     * @return \Org\Heigl\Hyphenator\Tokenizer\TokenRegistry
-     */
-    public function run($input)
-    {
-        if ($input instanceof TokenRegistry) {
-            // Tokenize a TokenRegistry
-            $f = clone($input);
-            foreach ($input as $token) {
-                if (! $token instanceof WordToken) {
-                    continue;
-                }
-                $newTokens = $this->tokenize($token->get());
-                if ($newTokens == array($token)) {
-                    continue;
-                }
-                $f->replace($token, $newTokens);
-            }
-
-            return $f ;
-        }
-
-        // Tokenize a simple string.
-        $array =  $this->tokenize($input);
-        $registry = new TokenRegistry();
-        foreach ($array as $item) {
-            $registry->add($item);
-        }
-
-        return $registry;
-    }
-
-    /**
-     * Split the given string into tokens using whitespace.
-     *
-     * Each whitespace is placed in a WhitespaceToken and everything else is
+     * Each punctuation is placed in a NonWordToken and everything else is
      * placed in a WordToken-Object
      *
      * @param \string $input The String to tokenize
      *
-     * @return Token
+     * @return Token[]
      */
-    private function tokenize($input)
+    protected function tokenize($input)
     {
         $tokens = array();
         $signs = '\\' . implode('\\', $this->tokens);
